move out Mock provider in a separate file
